+++ CHANGES.rst
@@ -1,6 +1,10 @@
 Changes
 =======
 
+In next release...
+
+- Fixed ZCA import fallbacks.
+
 1.0 (released 2009-11-01)
 -------------------------